Remove support of --ephemeral
This is actually useless in the case of running self-hosted runners on
Koyeb. It was first implemented because we wanted to have workers
running only one job, but it has a huge performance cost because of the
cold start. Instead, we'll update koyeb-github-runner-scheduler to
shutdown runners after a specified amount of time.